modules/queue/location/producer/requirements.txt

kafka-python
grpcio-tools